moved JetsonMotor class out of drive to extra motor class (drive is independent of motor type)
