+++ src/beachbot/manipulators/__init__.py
@@ -2,3 +2,5 @@ from .roarmm1 import RoArmM1
 from .drive import Motor, DifferentialDrive
 from .vrepmotorsim import VrepMotorSim
 from .vreproarmm1sim import VrepRoArmM1Sim
+
+from .jetsonmotor import JetsonMotor
